+++ src/avif.c
@@ -749,19 +749,8 @@ static avifBool overflowsInt32(int64_t x)
     return (x < INT32_MIN) || (x > INT32_MAX);
 }
 
-static avifBool avifCropRectIsValid(const avifCropRect * cropRect, uint32_t imageW, uint32_t imageH, avifPixelFormat yuvFormat, avifDiagnostics * diag)
-
+static avifBool avifCropRectIsValid(const avifCropRect * cropRect, uint32_t imageW, uint32_t imageH, avifDiagnostics * diag)
 {
-    // ISO/IEC 23000-22:2019/Amd. 2:2021, Section 7.3.6.7:
-    //   The clean aperture property is restricted according to the chroma
-    //   sampling format of the input image (4:4:4, 4:2:2:, 4:2:0, or 4:0:0) as
-    //   follows:
-    //   ...
-    //   - If chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0), the
-    //     leftmost pixel of the clean aperture shall be even numbers;
-    //   - If chroma is subsampled vertically (i.e., 4:2:0), the topmost line
-    //     of the clean aperture shall be even numbers.
-
     if ((cropRect->width == 0) || (cropRect->height == 0)) {
         avifDiagnosticsPrintf(diag, "[Strict] crop rect width and height must be nonzero");
         return AVIF_FALSE;
@@ -771,28 +760,16 @@ static avifBool avifCropRectIsValid(const avifCropRect * cropRect, uint32_t imag
         avifDiagnosticsPrintf(diag, "[Strict] crop rect is out of the image's bounds");
         return AVIF_FALSE;
     }
-
-    if ((yuvFormat == AVIF_PIXEL_FORMAT_YUV420) || (yuvFormat == AVIF_PIXEL_FORMAT_YUV422)) {
-        if ((cropRect->x % 2) != 0) {
-            avifDiagnosticsPrintf(diag, "[Strict] crop rect X offset must be even due to this image's YUV subsampling");
-            return AVIF_FALSE;
-        }
-    }
-    if (yuvFormat == AVIF_PIXEL_FORMAT_YUV420) {
-        if ((cropRect->y % 2) != 0) {
-            avifDiagnosticsPrintf(diag, "[Strict] crop rect Y offset must be even due to this image's YUV subsampling");
-            return AVIF_FALSE;
-        }
-    }
     return AVIF_TRUE;
 }
 
-avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
-                                             const avifCleanApertureBox * clap,
-                                             uint32_t imageW,
-                                             uint32_t imageH,
-                                             avifPixelFormat yuvFormat,
-                                             avifDiagnostics * diag)
+avifBool avifCropRectFromCleanApertureBox(avifCropRect * cropRect,
+                                          avifBool * upsampleBeforeCropping,
+                                          const avifCleanApertureBox * clap,
+                                          uint32_t imageW,
+                                          uint32_t imageH,
+                                          avifPixelFormat yuvFormat,
+                                          avifDiagnostics * diag)
 {
     avifDiagnosticsClearError(diag);
 
@@ -800,6 +777,16 @@ avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
     //   For horizOff and vertOff, D shall be strictly positive and N may be
     //   positive or negative. For cleanApertureWidth and cleanApertureHeight,
     //   N shall be positive and D shall be strictly positive.
+
+    // ISO/IEC 23000-22:2024, Section 7.3.6.7:
+    // The clean aperture property is restricted according to the chroma sampling format of the input image
+    // (4:4:4, 4:2:2, 4:2:0, or 4:0:0) as follows:
+    // - cleanApertureWidth and cleanApertureHeight shall be integers;
+    // - If any of the following conditions hold true, the image is first implicitly upsampled to 4:4:4:
+    //   - chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0) and cleanApertureWidth is odd
+    //   - chroma is subsampled horizontally (i.e., 4:2:2 and 4:2:0) and left-most pixel is on an odd position
+    //   - chroma is subsampled vertically (i.e., 4:2:0) and cleanApertureHeight is odd
+    //   - chroma is subsampled vertically (i.e., 4:2:0) and topmost line is on an odd position
 
     const int32_t widthN = (int32_t)clap->widthN;
     const int32_t widthD = (int32_t)clap->widthD;
@@ -898,7 +885,27 @@ avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
     cropRect->y = (uint32_t)(cropY.n / cropY.d);
     cropRect->width = (uint32_t)clapW;
     cropRect->height = (uint32_t)clapH;
-    return avifCropRectIsValid(cropRect, imageW, imageH, yuvFormat, diag);
+    if (!avifCropRectIsValid(cropRect, imageW, imageH, diag)) {
+        return AVIF_FALSE;
+    }
+
+    *upsampleBeforeCropping = ((yuvFormat == AVIF_PIXEL_FORMAT_YUV420 || yuvFormat == AVIF_PIXEL_FORMAT_YUV422) &&
+                               (cropRect->width % 2 || cropRect->x % 2)) ||
+                              (yuvFormat == AVIF_PIXEL_FORMAT_YUV420 && (cropRect->height % 2 || cropRect->y % 2));
+    return AVIF_TRUE;
+}
+
+avifBool avifCropRectConvertCleanApertureBox(avifCropRect * cropRect,
+                                             const avifCleanApertureBox * clap,
+                                             uint32_t imageW,
+                                             uint32_t imageH,
+                                             avifPixelFormat yuvFormat,
+                                             avifDiagnostics * diag)
+{
+    // Keep the same pre-deprecation behavior.
+    avifBool upsampleBeforeCropping;
+    return avifCropRectFromCleanApertureBox(cropRect, &upsampleBeforeCropping, clap, imageW, imageH, yuvFormat, diag) &&
+           !upsampleBeforeCropping;
 }
 
 avifBool avifCleanApertureBoxConvertCropRect(avifCleanApertureBox * clap,
@@ -908,9 +915,10 @@ avifBool avifCleanApertureBoxConvertCropRect(avifCleanApertureBox * clap,
                                              avifPixelFormat yuvFormat,
                                              avifDiagnostics * diag)
 {
+    (void)yuvFormat;
     avifDiagnosticsClearError(diag);
 
-    if (!avifCropRectIsValid(cropRect, imageW, imageH, yuvFormat, diag)) {
+    if (!avifCropRectIsValid(cropRect, imageW, imageH, diag)) {
         return AVIF_FALSE;
     }